Revert "move isWindowObject() function in utils"
This reverts commit d5c277632a91489e4ccb3c0a8ad74157ffa2db16.
